+++ storybook/storyLoader.js
@@ -4,10 +4,18 @@
 // https://github.com/elderfo/react-native-storybook-loader.git
 
 function loadStories() {
+  require('../src/components/ContactBraveBoxes/ContactBraveBoxes.stories')
   require('../src/components/HistoricAlert/HistoricAlert.stories')
+  require('../src/components/HomeScreenInstructions/HomeScreenInstructions.stories')
+  require('../src/components/InfoBox/InfoBox.stories')
 }
 
-const stories = ['../src/components/HistoricAlert/HistoricAlert.stories']
+const stories = [
+  '../src/components/ContactBraveBoxes/ContactBraveBoxes.stories',
+  '../src/components/HistoricAlert/HistoricAlert.stories',
+  '../src/components/HomeScreenInstructions/HomeScreenInstructions.stories',
+  '../src/components/InfoBox/InfoBox.stories',
+]
 
 module.exports = {
   loadStories,